Add 'fake' usrname & pword to url passed to git clone
The git credentials manager on Windows is a bit finicky and sometimes requires
providing credentials even when this is not strictly necessary. By
providing _some_ credentials in the url, we can force the credentials
manager to simply pass them along to GitHub which will ignore them due to
the repos being public.