Prepare dependencies without locking Cargo.lock
This doesn't pass --locked to Cargo, as that can interfere with upgrading the
lockfile format and cause spurious failures on Crater (since generate-lockfile,
which will get run in this case, will cause dependencies to get upgraded when
this isn't actually necessary). We would always regenerate the lockfile from
scratch if it was needed at all, so this isn't really changing anything behavior
wise.