Allow installing targets & components for CI toolchains
With this change, crater is able to successfully install alternative
targets for use with the `+target` modifier syntax.

While this works fine for crater, this doesn't work in the more general
case as the API allows you to call it multiple times with the assumption
being that doing so leaves the existing targets and components around
and only installs the newly requested target/component.
`rustup-toolchain-install-master` doesn't support this and instead
replaces the previous state of the toolchain with the newly requested
one.

As a result, calling `add_target("foo"); add_target("bar");` leaves only
the `bar` target installed.

Since installing CI toolchains is gated behind the unstable feature
flag, I think this is ok for now.