Add option to mount the source directory read-write
This adds the `Command::source_dir_mount_kind` method to provide the
ability to mount the source directory as read-write. This is needed for
the testing that we want to do in crater to test `cargo fix` and edition
migrations. This is needed because `cargo fix` inherently needs to
modify the source directory.